quasar: remove package
It's based on some old version of Vital which does not have all
fixes that were introduced in the current version of Vital.
(When I change current workspace in Hyprland, it crashes 0_0)
Also, in this fork, resizing and home-phoning were removed,
which is awful, because without auth, you can't even get basic shapes :|

The main reason to package this fork was a Renoise crash.
I'm going to try other free plugins, last resort is wine&serum.